Upload of current version of initial bins
This code shares many similarities with the original base code. It performs almost all the same analysis but introduces a temporal component where the analysis are repeated for the first, second and third five-second intervals of the data. These are then used as time-bins to analyze how the eye movement changes in the first 15 seconds.